Enemy spawning and some quick UI.

+++ Assets/_project/Scripts/Components/CollectableComponent.cs
@@ -5,6 +5,6 @@ namespace Game.Scripts
     [GenerateAuthoringComponent]
     public struct CollectableComponent : IComponentData
     {
-        public float Points;
+        public int Points;
     }
 }
